{"schema":"decodex/commit/1","summary":"Extract frozen selection image renderer","authority":"manual"}
